Made a new method of constructing EnumPlant values. Cleaned up a ton of unnecessary casts, missing @Overrides, changed private methods that could be static to either static or protected, and moved code in unnecessary else statements.

+++ src/main/java/genesis/world/biome/BiomeGenSwampRainforest.java
@@ -76,7 +76,7 @@ public class BiomeGenSwampRainforest extends BiomeGenBaseGenesis implements IEnt
 	@Override
 	public void genTerrainBlocks(World world, Random rand, ChunkPrimer p_180622_3_, int p_180622_4_, int p_180622_5_, double p_180622_6_)
 	{
-		double d1 = GRASS_COLOR_NOISE.func_151601_a((double)p_180622_4_ * 0.25D, (double)p_180622_5_ * 0.25D);
+		double d1 = GRASS_COLOR_NOISE.func_151601_a(p_180622_4_ * 0.25D, p_180622_5_ * 0.25D);
 		
 		if (d1 > -0.2D)
 		{
